extracted abstract base class
